pyproject.toml
[project]
name = "governance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.5",
    "opentimestamps-client>=0.7.2",
    "pymupdf>=1.27.2.3",
    "python-docx>=1.2.0",
    "pandas>=2.0",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "pyarrow>=24.0.0",
    "pyyaml>=6.0.3",
    "pypdf>=6.11.0",
    "shap>=0.49.1",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["wedge/tests"]
